Create el.json
Modern Greek translation (an on-going effort). Please note that some headings will be bilingual ie. Greek with some English terms, since the Greek language version of the regulatory terminology is somewhat confusing (my humble opinion).
Since this is work-in-progress I will be working on it every day for a few minutew-hours that I have available, and I will be uploading the updated version(s) of this excellent tool (Thank you CNIL!!!)